refactor(frontend): Move buttons logic to NavigationDrawerButtons
